+++ packages/ai/src/telemetry/index.ts
@@ -1,3 +1,3 @@
 export type { TelemetryOptions } from './telemetry-options';
 export type { TelemetryIntegration } from './telemetry-integration';
-export { registerTelemetry } from './telemetry-integration-registry';
+export { registerTelemetry } from './telemetry-registry';
